+++ src/Console/Application.php
@@ -35,24 +35,20 @@ declare(strict_types=1);
 
 namespace Infection\Console;
 
-use function extension_loaded;
+use function array_merge;
 use Infection\Command\ConfigureCommand;
 use Infection\Command\RunCommand;
-use Infection\Console\ConsoleOutput as InfectionConsoleOutput;
 use Infection\Container;
 use OutOfBoundsException;
 use PackageVersions\Versions;
-use const PHP_SAPI;
 use function preg_quote;
 use function Safe\preg_match;
+use function Safe\sprintf;
 use Symfony\Component\Console\Application as BaseApplication;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Formatter\OutputFormatterStyle;
-use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\ConsoleOutput;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Style\SymfonyStyle;
+use function trim;
 
 /**
  * @internal
@@ -69,14 +65,10 @@ final class Application extends BaseApplication
    / // __ \/ /_/ _ \/ ___/ __/ / __ \/ __ \
  _/ // / / / __/  __/ /__/ /_/ / /_/ / / / /
 /___/_/ /_/_/  \___/\___/\__/_/\____/_/ /_/
+
 ';
 
     private $container;
-
-    /**
-     * @var InfectionConsoleOutput
-     */
-    private $consoleOutput;
 
     public function __construct(Container $container)
     {
@@ -99,60 +91,34 @@ final class Application extends BaseApplication
         $this->setDefaultCommand('run');
     }
 
-    public function run(?InputInterface $input = null, ?OutputInterface $output = null): int
-    {
-        if ($input === null) {
-            $input = new ArgvInput();
-        }
-
-        if ($output === null) {
-            $output = new ConsoleOutput();
-        }
-
-        $this->consoleOutput = new InfectionConsoleOutput(new SymfonyStyle($input, $output));
-
-        if ($input->hasParameterOption(['--quiet', '-q'], true)) {
-            $output->setVerbosity(OutputInterface::VERBOSITY_QUIET);
-        }
-
-        $this->logRunningWithDebugger();
-
-        if (!$this->isAutoExitEnabled()) {
-            // When we're not in control of exit codes, that means it's the caller
-            // responsibility to disable xdebug if it isn't needed. As of writing
-            // that's only the case during E2E testing. Show a warning nevertheless.
-
-            $this->consoleOutput->logNotInControlOfExitCodes();
-
-            return parent::run($input, $output);
-        }
-
-        return parent::run($input, $output);
-    }
-
     public function getContainer(): Container
     {
         return $this->container;
     }
 
-    public function getConsoleOutput(): InfectionConsoleOutput
+    public function getLongVersion(): string
     {
-        return $this->consoleOutput;
+        return trim(sprintf(
+            '<info>%s</info> version <comment>%s</comment>',
+            $this->getName(),
+            $this->getVersion()
+        ));
     }
 
-    protected function doRunCommand(Command $command, InputInterface $input, OutputInterface $output)
+    public function getHelp(): string
     {
-        $output->writeln([self::LOGO, $this->getLongVersion()]);
-
-        return parent::doRunCommand($command, $input, $output);
+        return self::LOGO . parent::getHelp();
     }
 
     protected function getDefaultCommands()
     {
-        $commands = array_merge(parent::getDefaultCommands(), [
-            new ConfigureCommand(),
-            new RunCommand(),
-        ]);
+        $commands = array_merge(
+            parent::getDefaultCommands(),
+            [
+                new ConfigureCommand(),
+                new RunCommand(),
+            ]
+        );
 
         return $commands;
     }
@@ -174,16 +140,5 @@ final class Application extends BaseApplication
         $output->getFormatter()->setStyle('low', new OutputFormatterStyle('red', null, ['bold']));
         $output->getFormatter()->setStyle('medium', new OutputFormatterStyle('yellow', null, ['bold']));
         $output->getFormatter()->setStyle('high', new OutputFormatterStyle('green', null, ['bold']));
-    }
-
-    private function logRunningWithDebugger(): void
-    {
-        if (PHP_SAPI === 'phpdbg') {
-            $this->consoleOutput->logRunningWithDebugger(PHP_SAPI);
-        } elseif (extension_loaded('xdebug')) {
-            $this->consoleOutput->logRunningWithDebugger('Xdebug');
-        } elseif (extension_loaded('pcov')) {
-            $this->consoleOutput->logRunningWithDebugger('PCOV');
-        }
     }
 }
